I don't even know.....
Problem:
First of all the board in possibleMove should be dynamically allocated,
so as I mentioned last commit I had to change the Board function to
receive char** instead of char[][].
casting array to 1d array is done implicitly in C,
but apparently this is not the case in 2d.

We need to figure out a way for both dynamic and static arrays to
accepted by Board functions.