movePiece now works, scoring function upgraded
- a legal move now results in printing of the updated board, but then
sends the program into an infinite loop (because of minimax). Once we
sort out minimax everything should be alright.

- I moved the scoring function to Draughts.c so it could have access to
the global move lists and ask for their lengths.

-updatePossibleMove now updates both global move lists.
-fixed small bug with the quit command
-New iterator construction works OK as far as I can see.

Remaining issues:
-getPossibleJumps recursion

-verify minimax

-We should let the computer have the first move if he's playing the
white pieces. I tried to write it in but saw it was quite complicated.
How do you think it would be best to do this?